# [perf]: improve list retaining logic