I found that:

* `shouldRemove` function is not used in list-item opcode
*  it's `reset` function called in only one place and could be replaced with simple assignment

~In addition, it seems we do non-optimal thing here: `updateRef` is called if node is retained or moved, but it not necessary require tag update, adding `valueForRef` here to update tags that actually differs.~ And scope of `updateRef` usage reduced to `updateReferences` opcode function.

This PR a-bit conflicting with https://github.com/glimmerjs/glimmer-vm/pull/1529/files, but could be easily rebased.