@lifeart 2023-12-23
```
    Benchmark Results Summary    

duration phase no difference [-85ms to 63ms]
render1000Items1End phase no difference [-11ms to 3ms]
clearItems1End phase no difference [-1ms to 0ms]
render1000Items2End phase estimated improvement -7ms [-13ms to -1ms] OR -1.61% [-2.99% to -0.31%]
clearItems2End phase no difference [-1ms to 1ms]
render10000Items1End phase no difference [-58ms to 37ms]
clearItems3End phase no difference [-5ms to 5ms]
render1000Items3End phase no difference [-5ms to 12ms]
append1000Items1End phase no difference [-9ms to 5ms]
updateEvery10thItem1End phase no difference [-1ms to 0ms]
selectFirstRow1End phase no difference [0ms to 3ms]
selectSecondRow1End phase no difference [-2ms to 0ms]
removeFirstRow1End phase no difference [-1ms to 4ms]
removeSecondRow1End phase no difference [-2ms to 1ms]
swapRows1End phase no difference [-2ms to 2ms]
clearItems4End phase no difference [-1ms to 2ms]
paint phase no difference [0ms to 0ms]
```
[artifact-1.pdf](https://github.com/glimmerjs/glimmer-vm/files/13759357/artifact-1.pdf)